+++ R/datasummary_skim.R
@@ -154,6 +154,9 @@ datasummary_skim_numeric <- function(
   # output format
   sanitize_output(output)
 
+  # fake formula to indicate that we definitely use All() eventually
+  sanity_ds_data(All(data) ~ x + y, data)
+
   # draw histogram?
   if (histogram) {
 
@@ -357,11 +360,13 @@ datasummary_skim_categorical <- function(
   }
 
   if (!is.null(drop_too_many_levels)) {
-    warning(sprintf("These variables were omitted because they include more than 50 levels: %s.", paste(drop_too_many_levels, collapse=", ")))
+    warning(sprintf("These variables were omitted because they include more than 50 levels: %s.", paste(drop_too_many_levels, collapse=", ")),
+            call. = FALSE)
   }
 
   if (!is.null(drop_entirely_na)) {
-    warning(sprintf("These variables were omitted because they are entirely missing: %s.", paste(drop_entirely_na, collapse=", ")))
+    warning(sprintf("These variables were omitted because they are entirely missing: %s.", paste(drop_entirely_na, collapse=", ")),
+            call. = FALSE)
   }
 
   pctformat <- function(x) rounding(x, fmt)