+++ R/format_estimates.R
@@ -12,6 +12,7 @@ format_estimates <- function(
   fmt        = "%.3f",
   stars      = FALSE,
   group_name = NULL,
+  exponentiate = FALSE,
   ...) {
 
   # conf.int to glue
@@ -72,22 +73,30 @@ format_estimates <- function(
       estimate_glue[1] <- paste0(estimate_glue[1], "{stars}")
   }
 
+  # check if statistics are available -- too complicated for glue strings
+  statistic_nonglue <- c(estimate, statistic)
+  statistic_nonglue <- statistic_nonglue[!grepl("\\{", statistic_nonglue)]
+  statistic_nonglue[statistic_nonglue == "conf.int"] <- "conf.low"
+  for (s in statistic_nonglue) {
+    if (!s %in% colnames(est)) {
+      stop(sprintf("`%s` is not available. The `estimate` and `statistic` arguments must correspond to column names in the output of this command: `get_estimates(model)`", s))
+    }
+  }
 
-  ## I would like to check if the statistics are available and return an
-  ## informative message, but this does not seem possible for complicated glue
-  ## strings with functions:
-  ## 'p{ifelse(p.value <0.001, "Significant", "Not significant")'
+  # exponentiate
+  if (isTRUE(exponentiate)) {
 
-  ## # are statistics available? if not, display an informative error message
-  ## # check this after all custom statistics and stars are added
-  ## estimate_glue_strip <- regmatches(estimate_glue, gregexpr("\\{[^\\}]*\\}", estimate_glue))
-  ## estimate_glue_strip <- sort(unique(unlist(estimate_glue_strip)))
-  ## estimate_glue_strip <- gsub("\\{|\\}", "", estimate_glue_strip)
-  ## estimate_glue_strip <- setdiff(estimate_glue_strip, colnames(est))
-  ## if (length(estimate_glue_strip) > 0) {
-  ##   stop(sprintf("These estimates or statistics do not seem to be available: %s. You can use the `get_estimates` function to see which statistics are available.",
-  ##                paste(estimate_glue_strip, collapse = ", ")))
-  ## }
+    # standard error before transforming estimate
+    cols <- c("std.error", "estimate", "conf.low", "conf.high")
+    cols <- intersect(cols, colnames(est))
+    for (col in cols) {
+      if (col == "std.error") {
+        est[["std.error"]] <- exp(est[["estimate"]]) * est[["std.error"]]
+      } else if (col %in% c("estimate", "conf.low", "conf.high")) {
+        est[[col]] <- exp(est[[col]])
+      }
+    }
+  }
 
 
   ## round all 
@@ -104,7 +113,15 @@ format_estimates <- function(
   # extract estimates (there can be several)
   for (i in seq_along(estimate_glue)) {
     s <- estimate_glue[i]
-    est[[paste0("modelsummary_tmp", i)]] <- as.character(glue::glue_data(est, s))
+    # At this point, NAs are "". `glue_data(.na=NULL)` works with NAs
+    tmp <- est
+    tmp[tmp == ""] <- NA
+    if ("stars" %in% colnames(tmp)) {
+      tmp$stars[is.na(tmp$stars)] <- "" # otherwise NAs propagate when stars = FALSE in user-level call
+    }
+    tmp <- glue::glue_data(tmp, s, .na = NULL)
+    tmp[is.na(tmp)] <- ""
+    est[[paste0("modelsummary_tmp", i)]] <- as.character(tmp)
     # avoid empty parentheses for NAs
     est[[paste0("modelsummary_tmp", i)]][est[[s]] == ""] <- ""
   }
@@ -115,7 +132,8 @@ format_estimates <- function(
   } else if (!is.null(group_name)) {
     est[["group"]] <- ""
     warning(sprintf('Group name "%s" was not found in the extracted data. The "group" argument must be a column name in the data.frame produced by `get_estimates(model)`.  If you wish to combine models with and without grouped estimates, you will find examples on the modelsummary website:
-https://vincentarelbundock.github.io/modelsummary', group_name))
+https://vincentarelbundock.github.io/modelsummary', group_name),
+            call. = FALSE)
   } else {
     # cannot be NA because we need to merge
     est[["group"]] <- ""